properly cleanup when killed

+++ bulbs.js
@@ -44,10 +44,18 @@ module.exports = class Bulbs {
 }
 
 // when process  is kill, cleanup it's children
-process.once('SIGUSR2', function () {
-  forks.forEach((f) => {
-    process.kill(f.pid)
-  });
-  
-  process.kill(process.pid)
-});
+function exitHandler(options, exitCode) {
+    if (options.cleanup) { 
+      console.log('cleaning up spawned processes...');
+      forks.forEach((f) => {
+        process.kill(f.pid)
+      });
+    }
+    if (options.exit) process.exit();
+}
+
+process.on('exit', exitHandler.bind(null,{cleanup:true}));
+process.on('SIGINT', exitHandler.bind(null, {exit:true}));
+process.on('SIGUSR1', exitHandler.bind(null, {exit:true}));
+process.on('SIGUSR2', exitHandler.bind(null, {exit:true}));
+process.on('uncaughtException', exitHandler.bind(null, {exit:true}));
